Cleaners that have been modified -
To emulate the changes made to batteryClean.py. 

There are likely to be some errors as I am not able to effectively test them with the current parser.py. However, I will make some personal adjustments to the parser script to make it testable. 

There are still two more to do.